+++ src/alluvialdiagram/visualOptions.js
@@ -70,5 +70,14 @@ export const visualOptions = {
     default: 'Center',
   },
 
-  //@TODO: add colorscale based on source ID.
+  colorScale: {
+    type: 'colorScale',
+    label: 'Color scale',
+    dimension: 'source',
+    default: {
+      scaleType: 'ordinal',
+      interpolator: 'interpolateSpectral',
+    },
+    group: 'colors',
+  },
 }